// frontend/lib/api/generated/model/loopSessionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SpecResearch Loop
 * OpenAPI spec version: 0.1.0
 */
import type { CardResponse } from './cardResponse';
import type { LoopSessionResponseWorkingDraftNarrative } from './loopSessionResponseWorkingDraftNarrative';
import type { NodeHeadResponse } from './nodeHeadResponse';
import type { SpecVersionResponse } from './specVersionResponse';
import type { WorkflowNode } from './workflowNode';

export interface LoopSessionResponse {
  id: string;
  title: string | null;
  version: number;
  working_draft_node: WorkflowNode;
  working_draft_narrative: LoopSessionResponseWorkingDraftNarrative;
  node_heads: NodeHeadResponse[];
  cards: CardResponse[];
  produced_spec_version: SpecVersionResponse | null;
  valid_spec_version_id: string | null;
  created_at: string;
  updated_at: string;
}